Moved renderer out of jeweler dir
