+++ lib/bio-gem.rb
@@ -1,3 +1,4 @@
 #load Jeweler's mod for adapting it to our BioGem Project
+require 'bio-gem/generator/render'
 require 'bio-gem/mod/jeweler'
 require 'bio-gem/application'
